+++ rmk-macro/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Use 0x60000 as the default start addr for nRF52
+
 ## [0.2.0] - 2024-09-11
 
 ### Added